requirements.txt
ffmpeg==1.4
imageio==2.36.1
imageio-ffmpeg==0.5.1
numpy==2.2.1
pillow==11.1.0
psutil==6.1.1
python-dotenv==1.0.1
pytube==15.0.0
setuptools==75.8.0
six==1.17.0
whisper==1.1.10
youtube-dl==2021.12.17
yt-dlp==2025.1.15
